Moved functions outside of ml_worker module
