+++ docs/knowledge/catalogs/slicing-function-catalog/index.rst
@@ -3,13 +3,13 @@ Slicing functions
 
 - **Textual slicing function**
 
-  - :func:`~.giskard.ml_worker.testing.functions.slicing.short_comment_slicing_fn`
-  - :func:`~.giskard.ml_worker.testing.functions.slicing.keyword_lookup_slicing_fn`
-  - :func:`~.giskard.ml_worker.testing.functions.slicing.positive_sentiment_analysis`
-  - :func:`~.giskard.ml_worker.testing.functions.slicing.offensive_sentiment_analysis`
-  - :func:`~.giskard.ml_worker.testing.functions.slicing.irony_sentiment_analysis`
-  - :func:`~.giskard.ml_worker.testing.functions.slicing.hate_sentiment_analysis`
-  - :func:`~.giskard.ml_worker.testing.functions.slicing.emotion_sentiment_analysis`
+  - :func:`~.giskard.functions.slicing.short_comment_slicing_fn`
+  - :func:`~.giskard.functions.slicing.keyword_lookup_slicing_fn`
+  - :func:`~.giskard.functions.slicing.positive_sentiment_analysis`
+  - :func:`~.giskard.functions.slicing.offensive_sentiment_analysis`
+  - :func:`~.giskard.functions.slicing.irony_sentiment_analysis`
+  - :func:`~.giskard.functions.slicing.hate_sentiment_analysis`
+  - :func:`~.giskard.functions.slicing.emotion_sentiment_analysis`
 
 - **Numerical slicing function**
-  - :func:`~.giskard.ml_worker.testing.functions.slicing.outlier_filter`
+  - :func:`~.giskard.functions.slicing.outlier_filter`
