+++ docs/knowledge/catalogs/transformation-function-catalog/index.rst
@@ -3,7 +3,7 @@ Transformation functions
 
 - **Text transformation function**
 
-  - :func:`~.giskard.ml_worker.testing.functions.transformation.keyboard_typo_transformation`
-  - :func:`~.giskard.ml_worker.testing.functions.transformation.uppercase_transformation`
-  - :func:`~.giskard.ml_worker.testing.functions.transformation.lowercase_transformation`
-  - :func:`~.giskard.ml_worker.testing.functions.transformation.strip_punctuation`
+  - :func:`~.giskard.functions.transformation.keyboard_typo_transformation`
+  - :func:`~.giskard.functions.transformation.uppercase_transformation`
+  - :func:`~.giskard.functions.transformation.lowercase_transformation`
+  - :func:`~.giskard.functions.transformation.strip_punctuation`
