+++ docs/reference/slicing-functions/index.rst
@@ -12,15 +12,15 @@ Slicing functions
 Textual slicing
 ---------------
 
-.. autofunction:: giskard.ml_worker.testing.functions.slicing.short_comment_slicing_fn
-.. autofunction:: giskard.ml_worker.testing.functions.slicing.keyword_lookup_slicing_fn
-.. autofunction:: giskard.ml_worker.testing.functions.slicing.positive_sentiment_analysis
-.. autofunction:: giskard.ml_worker.testing.functions.slicing.offensive_sentiment_analysis
-.. autofunction:: giskard.ml_worker.testing.functions.slicing.irony_sentiment_analysis
-.. autofunction:: giskard.ml_worker.testing.functions.slicing.hate_sentiment_analysis
-.. autofunction:: giskard.ml_worker.testing.functions.slicing.emotion_sentiment_analysis
+.. autofunction:: giskard.functions.slicing.short_comment_slicing_fn
+.. autofunction:: giskard.functions.slicing.keyword_lookup_slicing_fn
+.. autofunction:: giskard.functions.slicing.positive_sentiment_analysis
+.. autofunction:: giskard.functions.slicing.offensive_sentiment_analysis
+.. autofunction:: giskard.functions.slicing.irony_sentiment_analysis
+.. autofunction:: giskard.functions.slicing.hate_sentiment_analysis
+.. autofunction:: giskard.functions.slicing.emotion_sentiment_analysis
 
 Numerical slicing functions
 ---------------------------
 
-.. autofunction:: giskard.ml_worker.testing.functions.slicing.outlier_filter
+.. autofunction:: giskard.functions.slicing.outlier_filter
