+++ giskard/functions/slicing.py
