--- giskard/ml_worker/testing/functions/__init__.py
@@ -1,4 +1,0 @@
-from giskard.ml_worker.testing.functions import slicing
-from giskard.ml_worker.testing.functions import transformation
-
-__all__ = ["slicing", "transformation"]
