+++ giskard/push/perturbation.py
@@ -8,16 +8,15 @@ Functions:
 - _check_after_perturbation: Check if perturbation changed the model's prediction.
 
 """
-from typing import Optional
-
 import hashlib
 
 import numpy as np
 import pandas as pd
+from typing import Optional
 
 from giskard.core.core import SupportedModelTypes
 from giskard.datasets.base import Dataset
-from giskard.ml_worker.testing.functions.transformation import add_value
+from giskard.functions.transformation import add_value
 from giskard.models.base.model import BaseModel
 from giskard.push.utils import (
     SupportedPerturbationType,
@@ -32,7 +31,6 @@ from giskard.scanner.robustness.text_transformations import (
     TextTypoTransformation,
     TextUppercase,
 )
-
 from ..push import PerturbationPush
 
 text_transformation_list = [
