+++ giskard/push/push_test_catalog/catalog.py
@@ -15,12 +15,12 @@ These test functions allow validating pushes and can be added to test suites.
 """
 from typing import Any, Optional
 
+from giskard.core.test_result import TestResult
 from giskard.datasets.base import Dataset
-from giskard.ml_worker.testing.functions.transformation import mad_transformation
+from giskard.functions.transformation import mad_transformation
+from giskard.models.base import BaseModel
 from giskard.registry.decorators import test
 from giskard.registry.slicing_function import SlicingFunction
-from giskard.core.test_result import TestResult
-from giskard.models.base import BaseModel
 from giskard.testing.tests.calibration import test_overconfidence_rate, test_underconfidence_rate
 from giskard.testing.tests.metamorphic import test_metamorphic_invariance
 from giskard.testing.tests.performance import _test_diff_prediction, test_f1, test_rmse
