+++ giskard/scanner/robustness/text_transformations.py
@@ -8,7 +8,7 @@ import pandas as pd
 
 from ...core.core import DatasetProcessFunctionMeta
 from ...datasets import Dataset
-from ...ml_worker.testing.functions.transformation import gruber
+from ...functions.transformation import gruber
 from ...registry.registry import get_object_uuid
 from ...registry.transformation_function import TransformationFunction
