+++ tests/test_push.py
@@ -5,9 +5,7 @@ import pandas as pd
 import pytest
 
 import giskard.push
-from giskard.ml_worker.testing.functions.transformation import mad_transformation
-from giskard.registry.giskard_test import GiskardTest
-from giskard.registry.slicing_function import slicing_function
+from giskard.functions.transformation import mad_transformation
 from giskard.push import Push
 from giskard.push.contribution import create_contribution_push
 from giskard.push.perturbation import create_perturbation_push
@@ -18,6 +16,8 @@ from giskard.push.utils import (
     coltype_to_supported_perturbation_type,
     slice_bounds_quartile,
 )
+from giskard.registry.giskard_test import GiskardTest
+from giskard.registry.slicing_function import slicing_function
 from giskard.slicing.slice import QueryBasedSliceFunction
 
 DATASETS = [
